+++ jdi-light/src/main/java/com/epam/jdi/light/elements/common/WindowsManager.java
@@ -11,7 +11,6 @@ import static com.epam.jdi.light.common.Exceptions.exception;
 import static com.epam.jdi.light.driver.WebDriverFactory.getDriver;
 import static com.epam.jdi.light.driver.WebDriverFactory.jsExecute;
 import static com.epam.jdi.light.settings.JDISettings.ELEMENT;
-import static com.epam.jdi.light.settings.WebSettings.logger;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 /**
@@ -24,16 +23,11 @@ public class WindowsManager {
     private static Safe<Boolean> newWindow = new Safe<>(() -> false);
 
     public static Set<String> getWindows() {
-        try {
-            Set<String> wHandles = getDriver().getWindowHandles();
-            if (windowHandles.get() != null && windowHandles.get().size() < wHandles.size())
-                newWindow.set(true);
-            windowHandles.set(wHandles);
-        }
-        catch (Throwable ignore) {
-            logger.error("Error during getWindows handles. Ignore error in case of mobile device");
-        }
-        return windowHandles.get();
+        Set<String> wHandles = getDriver().getWindowHandles();
+        if (windowHandles.get() != null && windowHandles.get().size() < wHandles.size())
+            newWindow.set(true);
+        windowHandles.set(wHandles);
+        return wHandles;
     }
 
     /**